# cmd/server/main.go
package main

import (
	"bytes"
	"fmt"
	"html/template"
	"io"
	"log"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/gorilla/mux"
	"github.com/rs/cors"

	"warehousecore/config"
	"warehousecore/internal/handlers"
	"warehousecore/internal/led"
	"warehousecore/internal/middleware"
	"warehousecore/internal/models"
	"warehousecore/internal/repository"
	"warehousecore/internal/services"
)

var brandingService *services.CompanyBrandingService

// spaHandler serves the SPA and falls back to index.html for client-side routes
func spaHandler(w http.ResponseWriter, r *http.Request) {
	// Build file path
	path := "./web/dist" + r.URL.Path

	// Check if file exists
	fileInfo, err := os.Stat(path)
	if os.IsNotExist(err) {
		// File doesn't exist - serve index.html with injected config
		serveIndexWithConfig(w, r)
		return
	}

	// If path is a directory (not a file), serve index.html for SPA routing
	if err == nil && fileInfo.IsDir() {
		serveIndexWithConfig(w, r)
		return
	}

	// File exists and is not a directory - serve it
	http.FileServer(http.Dir("./web/dist")).ServeHTTP(w, r)
}

// serveIndexWithConfig injects runtime configuration into index.html
func serveIndexWithConfig(w http.ResponseWriter, r *http.Request) {
	// Read the index.html file
	file, err := os.Open("./web/dist/index.html")
	if err != nil {
		http.Error(w, "Could not open index.html", http.StatusInternalServerError)
		return
	}
	defer file.Close()

	content, err := io.ReadAll(file)
	if err != nil {
		http.Error(w, "Could not read index.html", http.StatusInternalServerError)
		return
	}

	// Get domain configuration from environment variables
	// These should be just the domain (e.g., "rent.server-nt.de")
	// without protocol or port - the frontend will add the protocol
	rentalCoreDomain := os.Getenv("RENTALCORE_DOMAIN")
	warehouseCoreDomain := os.Getenv("WAREHOUSECORE_DOMAIN")

	// Resolve company branding
	companyName := "RentalCore"
	if brandingService != nil {
		companyName = brandingService.CompanyName()
	}

	// Resolve currency symbol
	currencySymbol := services.GetCurrencySymbol()

	// Create config injection script
	configScript := fmt.Sprintf(
		`<script>window.__APP_CONFIG__={rentalCoreDomain:"%s",warehouseCoreDomain:"%s",companyName:"%s",currencySymbol:"%s"};</script>`,
		template.JSEscapeString(rentalCoreDomain),
		template.JSEscapeString(warehouseCoreDomain),
		template.JSEscapeString(companyName),
		template.JSEscapeString(currencySymbol),
	)

	// Inject the script before </head>
	modifiedContent := bytes.Replace(content, []byte("</head>"), []byte(configScript+"</head>"), 1)

	// Serve the modified content
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	w.Write(modifiedContent)
}

func main() {
	// Load configuration
	cfg, err := config.Load()
	if err != nil {
		log.Fatalf("Failed to load config: %v", err)
	}

	// Initialize database
	if err := repository.InitDatabase(cfg); err != nil {
		log.Fatalf("Failed to initialize database: %v", err)
	}
	defer repository.CloseDatabase()

	brandingService = services.NewCompanyBrandingService(repository.GetDB())

	// Initialize LED service
	log.Println("[LED] Initializing LED service...")
	_ = led.GetService() // Initialize singleton at startup
	log.Println("[LED] LED service initialization complete")

	led.RegisterControllerHeartbeatHandler(func(identifier string, payload *models.LEDControllerHeartbeat) error {
		service := services.NewLEDControllerService()
		_, err := service.RecordHeartbeat(identifier, payload)
		return err
	})

	controllerListener := led.NewControllerListener()
	if controllerListener != nil {
		controllerListener.Start()
	}

	// Ensure auto-admin assignment based on ENV (ADMIN_NAME_MATCH)
	go func() {
		// Run asynchronously to not block startup
		r := services.NewRBACService()
		if err := r.EnsureAutoAdminFromEnv(); err != nil {
			log.Printf("[RBAC] Auto-admin assignment failed: %v", err)
		}
	}()

	// Bootstrap Eventory config from environment variables (if any and not yet stored)
	services.BootstrapEventoryFromEnv()

	// Start Eventory background scheduler (no-op if sync_interval_minutes == 0)
	services.GetEventoryScheduler().Reset()

	// Setup router
	router := mux.NewRouter()

	// API v1 routes
	api := router.PathPrefix("/api/v1").Subrouter()

	// Auth endpoints (public - no auth required)
	api.HandleFunc("/auth/login", handlers.Login).Methods("POST")
	api.HandleFunc("/auth/logout", handlers.Logout).Methods("POST")

	// Health check (public)
	api.HandleFunc("/health", handlers.HealthCheck).Methods("GET")

	// App config (public - returns live branding info)
	api.HandleFunc("/config", func(w http.ResponseWriter, r *http.Request) {
		name := "RentalCore"
		if brandingService != nil {
			name = brandingService.CompanyName()
		}
		currencySymbol := services.GetCurrencySymbol()
		w.Header().Set("Content-Type", "application/json")
		w.Header().Set("Cache-Control", "no-store")
		fmt.Fprintf(w, `{"companyName":%q,"currencySymbol":%q}`, name, currencySymbol)
	}).Methods("GET")

	// Public product pictures (must be accessible without headers for IMG tags)
	api.HandleFunc("/public/products/{id}/pictures/{filename}", handlers.DownloadProductPicture).Methods("GET", "HEAD")

	// Public website feeds (no authentication required)
	api.HandleFunc("/public/products", handlers.GetWebsiteProducts).Methods("GET")
	api.HandleFunc("/public/packages", handlers.GetWebsitePackages).Methods("GET")

	// Protected routes - apply auth middleware
	protected := api.PathPrefix("").Subrouter()
	protected.Use(middleware.AuthMiddleware)

	// Auth status endpoint (requires authentication)
	protected.HandleFunc("/auth/me", handlers.GetCurrentUser).Methods("GET")
	protected.HandleFunc("/auth/change-password", handlers.ChangePassword).Methods("POST")

	// Scan endpoints (CRITICAL - core functionality)
	protected.HandleFunc("/scans", handlers.HandleScan).Methods("POST")
	protected.HandleFunc("/scans/history", handlers.GetScanHistory).Methods("GET")

	// Device endpoints
	api.HandleFunc("/devices", handlers.GetDevices).Methods("GET")
	api.HandleFunc("/devices/tree", handlers.GetDeviceTree).Methods("GET")
	api.HandleFunc("/devices/{id}", handlers.GetDevice).Methods("GET")
	api.HandleFunc("/devices/{id}/status", handlers.UpdateDeviceStatus).Methods("PUT")
	api.HandleFunc("/devices/{id}/movements", handlers.GetDeviceMovements).Methods("GET")

	// Zone endpoints
	api.HandleFunc("/zones", handlers.GetZones).Methods("GET")
	api.HandleFunc("/zones", handlers.CreateZone).Methods("POST")
	api.HandleFunc("/zones/scan", handlers.GetZoneByBarcode).Methods("GET")             // Zone barcode lookup
	api.HandleFunc("/zones/{id}/devices", handlers.GetZoneDevices).Methods("GET")       // Must be before /zones/{id}
	api.HandleFunc("/zones/{id}/devices", handlers.AssignDevicesToZone).Methods("POST") // Assign devices to zone
	api.HandleFunc("/zones/{id}/products", handlers.GetZoneProducts).Methods("GET")     // Get products in zone
	api.HandleFunc("/zones/{id}", handlers.GetZone).Methods("GET")
	api.HandleFunc("/zones/{id}", handlers.UpdateZone).Methods("PUT")
	api.HandleFunc("/zones/{id}", handlers.DeleteZone).Methods("DELETE")
	api.HandleFunc("/zone-types", handlers.GetZoneTypes).Methods("GET")

	// Job endpoints
	api.HandleFunc("/jobs", handlers.GetJobs).Methods("GET")
	api.HandleFunc("/jobs/{id}", handlers.GetJobSummary).Methods("GET")
	api.HandleFunc("/jobs/{id}/complete", handlers.CompleteJob).Methods("POST")

	// Public rental equipment endpoint (for RentalCore integration)
	api.HandleFunc("/rental-equipment", handlers.GetRentalEquipment).Methods("GET")
	api.HandleFunc("/rental-equipment/{id}", handlers.GetRentalEquipmentByID).Methods("GET")

	// Inventory endpoints
	api.HandleFunc("/inventory/low-stock", handlers.GetLowStockAlerts).Methods("GET")

	// Case endpoints
	api.HandleFunc("/cases/scan", handlers.GetCaseByScan).Methods("GET")
	api.HandleFunc("/cases", handlers.GetCases).Methods("GET")
	api.HandleFunc("/cases", handlers.CreateCase).Methods("POST")
	api.HandleFunc("/cases/{id}", handlers.GetCase).Methods("GET")
	api.HandleFunc("/cases/{id}", handlers.UpdateCase).Methods("PUT")
	api.HandleFunc("/cases/{id}", handlers.DeleteCase).Methods("DELETE")
	api.HandleFunc("/cases/{id}/contents", handlers.GetCaseContents).Methods("GET")
	api.HandleFunc("/cases/{id}/devices", handlers.AddDevicesToCase).Methods("POST")
	api.HandleFunc("/cases/{id}/devices/{device_id}", handlers.RemoveDeviceFromCase).Methods("DELETE")

	// Maintenance endpoints
	api.HandleFunc("/defects", handlers.GetDefects).Methods("GET")
	api.HandleFunc("/defects", handlers.CreateDefect).Methods("POST")
	api.HandleFunc("/defects/{id}", handlers.UpdateDefect).Methods("PUT")
	api.HandleFunc("/maintenance/inspections", handlers.GetInspections).Methods("GET")
	api.HandleFunc("/maintenance/stats", handlers.GetMaintenanceStats).Methods("GET")

	// Dashboard/stats
	api.HandleFunc("/dashboard/stats", handlers.GetDashboardStats).Methods("GET")
	api.HandleFunc("/movements", handlers.GetMovements).Methods("GET")

	// LED control endpoints
	api.HandleFunc("/led/status", handlers.GetLEDStatus).Methods("GET")
	api.HandleFunc("/led/highlight", handlers.HighlightJobBins).Methods("POST")
	api.HandleFunc("/led/clear", handlers.ClearLEDs).Methods("POST")
	api.HandleFunc("/led/identify", handlers.IdentifyLEDs).Methods("POST")
	api.HandleFunc("/led/test", handlers.TestBin).Methods("POST")
	api.HandleFunc("/led/locate", handlers.LocateBin).Methods("POST")

	// Label generation endpoints
	api.HandleFunc("/labels/qrcode", handlers.GenerateQRCode).Methods("POST")
	api.HandleFunc("/labels/barcode", handlers.GenerateBarcode).Methods("POST")
	api.HandleFunc("/labels/templates", handlers.GetLabelTemplates).Methods("GET")
	api.HandleFunc("/labels/templates", handlers.CreateLabelTemplate).Methods("POST")
	api.HandleFunc("/labels/templates/{id}", handlers.GetLabelTemplate).Methods("GET")
	api.HandleFunc("/labels/templates/{id}", handlers.UpdateLabelTemplate).Methods("PUT")
	api.HandleFunc("/labels/templates/{id}", handlers.DeleteLabelTemplate).Methods("DELETE")
	api.HandleFunc("/labels/device/{device_id}", handlers.GenerateDeviceLabel).Methods("POST")
	api.HandleFunc("/labels/case/{case_id}", handlers.GenerateCaseLabel).Methods("POST")
	api.HandleFunc("/labels/zone/{zone_id}", handlers.GenerateZoneLabel).Methods("POST")
	api.HandleFunc("/labels/save", handlers.SaveDeviceLabel).Methods("POST")
	api.HandleFunc("/labels/save-case", handlers.SaveCaseLabel).Methods("POST")
	api.HandleFunc("/labels/save-zone", handlers.SaveZoneLabel).Methods("POST")

	// Admin routes (RBAC protected)
	// Read-only admin routes (admin or manager)
	adminRead := api.PathPrefix("/admin").Subrouter()
	adminRead.Use(middleware.AuthMiddleware)
	adminRead.Use(middleware.RequireAdminOrManager)
	adminRead.HandleFunc("/zone-types", handlers.GetZoneTypes).Methods("GET")
	adminRead.HandleFunc("/zone-types/{id}", handlers.GetZoneType).Methods("GET")
	adminRead.HandleFunc("/led/single-bin-default", handlers.GetLEDSingleBinDefault).Methods("GET")
	adminRead.HandleFunc("/led/job-highlights", handlers.GetLEDJobHighlightSettings).Methods("GET")
	adminRead.HandleFunc("/led/mapping", handlers.GetLEDMapping).Methods("GET")
	adminRead.HandleFunc("/led/controllers", handlers.ListLEDControllers).Methods("GET")
	adminRead.HandleFunc("/led/controllers", handlers.ListLEDControllers).Methods("GET")
	adminRead.HandleFunc("/roles", handlers.GetRoles).Methods("GET")
	adminRead.HandleFunc("/users", handlers.GetUsersWithRoles).Methods("GET")
	adminRead.HandleFunc("/users/{id}/roles", handlers.GetUserRoles).Methods("GET")
	adminRead.HandleFunc("/categories", handlers.GetCategories).Methods("GET")
	adminRead.HandleFunc("/subcategories", handlers.GetSubcategories).Methods("GET")
	adminRead.HandleFunc("/subbiercategories", handlers.GetSubbiercategories).Methods("GET")
	adminRead.HandleFunc("/count-types", handlers.ListCountTypes).Methods("GET")
	adminRead.HandleFunc("/brands", handlers.GetBrands).Methods("GET")
	adminRead.HandleFunc("/manufacturers", handlers.GetManufacturers).Methods("GET")
	adminRead.HandleFunc("/products", handlers.GetProducts).Methods("GET")
	adminRead.HandleFunc("/products/{id}", handlers.GetProduct).Methods("GET")
	adminRead.HandleFunc("/products/{id}/pictures", handlers.GetProductPictures).Methods("GET")
	adminRead.HandleFunc("/products/{id}/pictures/{filename}", handlers.DownloadProductPicture).Methods("GET", "HEAD")
	adminRead.HandleFunc("/product-packages", handlers.GetProductPackages).Methods("GET")
	adminRead.HandleFunc("/product-packages/{id}", handlers.GetProductPackage).Methods("GET")
	adminRead.HandleFunc("/rental-equipment", handlers.GetRentalEquipment).Methods("GET")
	adminRead.HandleFunc("/rental-equipment/suppliers", handlers.GetRentalEquipmentSuppliers).Methods("GET")
	adminRead.HandleFunc("/rental-equipment/{id}", handlers.GetRentalEquipmentByID).Methods("GET")
	adminRead.HandleFunc("/api-keys", handlers.ListAPIKeys).Methods("GET")

	// Admin-only routes (write operations)
	admin := api.PathPrefix("/admin").Subrouter()
	admin.Use(middleware.AuthMiddleware)
	admin.Use(middleware.RequireAdmin)
	admin.HandleFunc("/zone-types", handlers.CreateZoneType).Methods("POST")
	admin.HandleFunc("/zone-types/{id}", handlers.UpdateZoneType).Methods("PUT")
	admin.HandleFunc("/zone-types/{id}", handlers.DeleteZoneType).Methods("DELETE")
	admin.HandleFunc("/led/single-bin-default", handlers.UpdateLEDSingleBinDefault).Methods("PUT")
	admin.HandleFunc("/led/job-highlights", handlers.UpdateLEDJobHighlightSettings).Methods("PUT")
	admin.HandleFunc("/led/mapping", handlers.UpdateLEDMapping).Methods("PUT")
	admin.HandleFunc("/led/mapping/validate", handlers.ValidateLEDMapping).Methods("POST")
	admin.HandleFunc("/led/preview", handlers.PreviewLEDSettings).Methods("POST")
	admin.HandleFunc("/led/controllers/{id}", handlers.UpdateLEDController).Methods("PUT")
	admin.HandleFunc("/led/controllers", handlers.CreateLEDController).Methods("POST")
	admin.HandleFunc("/led/controllers/{id}", handlers.UpdateLEDController).Methods("PUT")
	admin.HandleFunc("/led/controllers/{id}/configure", handlers.ConfigureLEDController).Methods("POST")
	admin.HandleFunc("/led/controllers/{id}/restart", handlers.RestartLEDController).Methods("POST")
	admin.HandleFunc("/led/controllers/{id}", handlers.DeleteLEDController).Methods("DELETE")
	admin.HandleFunc("/users/{id}/roles", handlers.UpdateUserRoles).Methods("PUT")
	admin.HandleFunc("/categories", handlers.CreateCategory).Methods("POST")
	admin.HandleFunc("/categories/{id}", handlers.UpdateCategory).Methods("PUT")
	admin.HandleFunc("/categories/{id}", handlers.DeleteCategory).Methods("DELETE")
	admin.HandleFunc("/subcategories", handlers.CreateSubcategory).Methods("POST")
	admin.HandleFunc("/subcategories/{id}", handlers.UpdateSubcategory).Methods("PUT")
	admin.HandleFunc("/subcategories/{id}", handlers.DeleteSubcategory).Methods("DELETE")
	admin.HandleFunc("/subbiercategories", handlers.CreateSubbiercategory).Methods("POST")
	admin.HandleFunc("/subbiercategories/{id}", handlers.UpdateSubbiercategory).Methods("PUT")
	admin.HandleFunc("/subbiercategories/{id}", handlers.DeleteSubbiercategory).Methods("DELETE")
	admin.HandleFunc("/count-types", handlers.CreateCountType).Methods("POST")
	admin.HandleFunc("/count-types/{id}", handlers.UpdateCountType).Methods("PUT")
	admin.HandleFunc("/count-types/{id}", handlers.DeleteCountType).Methods("DELETE")
	admin.HandleFunc("/brands", handlers.CreateBrand).Methods("POST")
	admin.HandleFunc("/brands/{id}", handlers.UpdateBrand).Methods("PUT")
	admin.HandleFunc("/brands/{id}", handlers.DeleteBrand).Methods("DELETE")
	admin.HandleFunc("/manufacturers", handlers.CreateManufacturer).Methods("POST")
	admin.HandleFunc("/manufacturers/{id}", handlers.UpdateManufacturer).Methods("PUT")
	admin.HandleFunc("/manufacturers/{id}", handlers.DeleteManufacturer).Methods("DELETE")
	admin.HandleFunc("/products/bulk-delete", handlers.BulkDeleteProducts).Methods("POST")
	admin.HandleFunc("/products/bulk-update", handlers.BulkUpdateProducts).Methods("PUT")
	admin.HandleFunc("/products", handlers.CreateProduct).Methods("POST")
	admin.HandleFunc("/products/{id}", handlers.UpdateProduct).Methods("PUT")
	admin.HandleFunc("/products/{id}", handlers.DeleteProduct).Methods("DELETE")
	admin.HandleFunc("/products/{id}/website", handlers.UpdateProductWebsite).Methods("PUT")
	admin.HandleFunc("/products/{id}/pictures", handlers.UploadProductPictures).Methods("POST")
	admin.HandleFunc("/products/{id}/pictures/{filename}", handlers.DeleteProductPicture).Methods("DELETE")
	admin.HandleFunc("/products/{id}/devices", handlers.GetProductDevices).Methods("GET")
	admin.HandleFunc("/products/{id}/devices", handlers.CreateDevicesForProduct).Methods("POST")
	admin.HandleFunc("/products/{id}/dependencies", handlers.GetProductDependencies).Methods("GET")
	admin.HandleFunc("/products/{id}/dependencies", handlers.CreateProductDependency).Methods("POST")
	admin.HandleFunc("/products/{id}/dependencies/{dep_id}", handlers.DeleteProductDependency).Methods("DELETE")
	admin.HandleFunc("/products/{id}/convert-to-case", handlers.ConvertProductToCase).Methods("POST")
	admin.HandleFunc("/products/{id}/convert-to-cable", handlers.ConvertProductToCable).Methods("POST")
	admin.HandleFunc("/product-packages", handlers.CreateProductPackage).Methods("POST")
	admin.HandleFunc("/product-packages/{id}", handlers.UpdateProductPackage).Methods("PUT")
	admin.HandleFunc("/product-packages/{id}", handlers.DeleteProductPackage).Methods("DELETE")
	admin.HandleFunc("/product-packages/{id}/items", handlers.AddItemToPackage).Methods("POST")
	admin.HandleFunc("/product-packages/{package_id}/items/{item_id}", handlers.RemoveItemFromPackage).Methods("DELETE")
	admin.HandleFunc("/rental-equipment", handlers.CreateRentalEquipment).Methods("POST")
	admin.HandleFunc("/rental-equipment/{id}", handlers.UpdateRentalEquipment).Methods("PUT")
	admin.HandleFunc("/rental-equipment/{id}", handlers.DeleteRentalEquipment).Methods("DELETE")
	admin.HandleFunc("/api-keys", handlers.CreateAPIKey).Methods("POST")
	admin.HandleFunc("/api-keys/{id}/status", handlers.UpdateAPIKeyStatus).Methods("PUT")
	admin.HandleFunc("/api-keys/{id}", handlers.DeleteAPIKey).Methods("DELETE")

	// Cable admin endpoints
	admin.HandleFunc("/cables", handlers.GetAllCables).Methods("GET")
	admin.HandleFunc("/cables", handlers.CreateCable).Methods("POST")
	admin.HandleFunc("/cables/{id}", handlers.GetCable).Methods("GET")
	admin.HandleFunc("/cables/{id}", handlers.UpdateCable).Methods("PUT")
	admin.HandleFunc("/cables/{id}", handlers.DeleteCable).Methods("DELETE")
	admin.HandleFunc("/cables/{id}/devices", handlers.GetCableDevices).Methods("GET")
	admin.HandleFunc("/cables/{id}/devices", handlers.CreateDevicesForCable).Methods("POST")
	admin.HandleFunc("/cable-connectors", handlers.GetCableConnectors).Methods("GET")
	admin.HandleFunc("/cable-types", handlers.GetCableTypes).Methods("GET")
	admin.HandleFunc("/devices-list", handlers.GetAllDevicesAdmin).Methods("GET")
	admin.HandleFunc("/devices/bulk-delete", handlers.BulkDeleteDevices).Methods("POST")
	admin.HandleFunc("/devices/bulk-update", handlers.BulkUpdateDevices).Methods("PUT")
	admin.HandleFunc("/devices", handlers.CreateDevice).Methods("POST")
	admin.HandleFunc("/devices/{id}", handlers.GetDeviceAdmin).Methods("GET")
	admin.HandleFunc("/devices/{id}", handlers.UpdateDevice).Methods("PUT")
	admin.HandleFunc("/devices/{id}", handlers.DeleteDevice).Methods("DELETE")
	admin.HandleFunc("/devices/{id}/qr", handlers.GenerateDeviceQR).Methods("GET")
	admin.HandleFunc("/devices/{id}/barcode", handlers.GenerateDeviceBarcode).Methods("GET")
	admin.HandleFunc("/api-limits", handlers.GetAPILimits).Methods("GET")
	admin.HandleFunc("/api-limits", handlers.UpdateAPILimits).Methods("PUT")
	admin.HandleFunc("/currency", handlers.GetCurrencySettings).Methods("GET")
	admin.HandleFunc("/currency", handlers.UpdateCurrencySettings).Methods("PUT")

	// Eventory integration endpoints (admin-only write + read)
	admin.HandleFunc("/eventory/settings", handlers.GetEventorySettings).Methods("GET")
	admin.HandleFunc("/eventory/settings", handlers.UpdateEventorySettings).Methods("PUT")
	admin.HandleFunc("/eventory/products", handlers.GetEventoryProducts).Methods("GET")
	admin.HandleFunc("/eventory/sync", handlers.SyncEventoryProducts).Methods("POST")
	admin.HandleFunc("/eventory/credential-key", handlers.GetEventoryCredentialKeyStatus).Methods("GET")
	admin.HandleFunc("/eventory/credential-key", handlers.UpdateEventoryCredentialKey).Methods("PUT")
	admin.HandleFunc("/eventory/credential-key/generate", handlers.GenerateEventoryCredentialKey).Methods("POST")

	// CSV Export endpoints (read-only, admin or manager)
	adminRead.HandleFunc("/export/{type}", handlers.ExportCSV).Methods("GET")

	// Profile endpoints (authenticated users)
	protected.HandleFunc("/profile/me", handlers.GetMyProfile).Methods("GET")
	protected.HandleFunc("/profile/me", handlers.UpdateMyProfile).Methods("PUT")
	protected.HandleFunc("/product-packages/alias-map", handlers.GetProductPackageAliasMap).Methods("GET")

	// Service API routes (service-to-service, any active API key)
	serviceAPI := api.PathPrefix("/service").Subrouter()
	serviceAPI.Use(middleware.APIKeyMiddleware)
	serviceAPI.HandleFunc("/cables", handlers.GetAllCables).Methods("GET")
	serviceAPI.HandleFunc("/cables/{id}", handlers.GetCable).Methods("GET")
	serviceAPI.HandleFunc("/devices/{id}", handlers.GetDevice).Methods("GET")

	// Apply middleware
	api.Use(middleware.Logger)
	api.Use(middleware.RecoveryMiddleware)
	// Docs endpoints - always available but require authentication
	docsRouter := router.PathPrefix("").Subrouter()
	docsRouter.Use(middleware.AuthMiddleware)
	registerDynamicDocs(docsRouter, router)

	// Serve static frontend files with SPA fallback
	router.PathPrefix("/").HandlerFunc(spaHandler)

	// CORS setup
	c := cors.New(cors.Options{
		AllowedOrigins:   cfg.CORS.AllowedOrigins,
		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Content-Type", "Authorization", "X-API-Key"},
		AllowCredentials: true,
	})

	handler := c.Handler(router)

	// Server configuration
	server := &http.Server{
		Addr:         cfg.Server.Host + ":" + cfg.Server.Port,
		Handler:      handler,
		ReadTimeout:  15 * time.Second,
		WriteTimeout: 15 * time.Second,
		IdleTimeout:  60 * time.Second,
	}

	// Start server in goroutine
	go func() {
		log.Printf("WarehouseCore server starting on %s:%s", cfg.Server.Host, cfg.Server.Port)
		log.Printf("Environment: %s", cfg.App.Environment)
		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Fatalf("Server failed: %v", err)
		}
	}()

	// Graceful shutdown
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
	<-quit

	log.Println("Shutting down server...")
	if controllerListener != nil {
		controllerListener.Close()
	}
	services.GetEventoryScheduler().Stop()
	repository.CloseDatabase()
	log.Println("Server stopped")
}